Updates LinkedIn Sales Navigator settings to cover solution merge
With the merged solutions between V1 and V2, the installation process no
longer involves an actual installation. Instead there's a way to quickly
get enable the forms as an administrator, but the packages themselves
remain the same functionally.

This updates the docs to provide information on how to enable these
default forms as an administrator, as well as how to disable them. The
documentation around the older AppSource packages has removed given that
these packages will be removed from there and included out of the box.